+++ storylets/CHANGELOG.md
@@ -3,3 +3,6 @@
 * ''v1.2 — 2023-10-10'' Added the `pulls` condition
 * ''v1.3 - 2023-10-27'' BUG FIX: sort negative priority correctly
 * ''v1.4 - 2023-11-15'' BUG FIX: don't initialise a store if it already exists
+* ''v1.5 - 2024-08-29'' Added the `twinescript (tws)` condition type. Added the `<<storyletinclude>>` macro. 
+                        Changed syntax for `<<storyletscan>>`. Refactor of the `Sequence` class.
+                        Added new syntax for `<<storylet>>` macro.